requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.9.2
allennlp==2.10.1
amrlib==0.7.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
autopage==0.5.1
Babel==2.12.1
base58==2.1.1
bertviz==1.4.0
blis==0.7.9
boto3==1.26.80
botocore==1.29.80
brotlipy==0.7.0
cached-path==1.3.3
cachetools==5.3.0
catalogue==2.0.8
cfgv==3.3.1
cliff==4.1.0
cmaes==0.9.1
cmd2==2.4.3
colorama==0.4.6
colorlog==6.7.0
confection==0.0.4
cymem==2.0.7
datasets==2.9.0
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.0
emoji==2.2.0
exceptiongroup==1.1.0
fairscale==0.4.13
filelock==3.9.0
fire==0.5.0
fonttools==4.25.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.0
hydra-colorlog==1.2.0
hydra-core==1.3.1
hydra-optuna-sweeper==1.2.0
identify==2.5.17
importlib-metadata==6.0.0
iniconfig==2.0.0
isoduration==20.11.0
jmespath==1.0.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_server_fileid==0.8.0
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab_server==2.20.0
langcodes==3.3.0
lightning-utilities==0.3.0
lion-pytorch==0.0.7
llvmlite==0.39.1
lmdb==1.4.0
Mako==1.2.4
Markdown==3.4.1
mkl-fft==1.3.1
mkl-service==2.4.0
more-itertools==9.1.0
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
murmurhash==1.0.9
mypy-extensions==0.4.3
nashpy==0.0.36
networkx==3.0
nodeenv==1.7.0
oauthlib==3.2.2
omegaconf==2.3.0
optuna==2.10.1
pandas==1.5.2
pathtools==0.1.2
pathy==0.10.1
pbr==5.11.1
Penman==1.2.2
Pillow==9.3.0
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
pre-commit==3.0.4
preshed==3.0.8
prettytable==3.6.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pydot==1.4.2
pyperclip==1.8.2
PyQt5-sip==12.11.0
pyrootutils==1.0.4
pytest==7.2.1
python-dotenv==0.21.1
python-json-logger==2.0.7
pytorch-lightning==1.8.3
PyYAML==6.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rotary-embedding-torch==0.2.1
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.3.1
scipy==1.9.3
sentencepiece==0.1.95
sentry-sdk==1.14.0
setproctitle==1.3.2
smart-open==6.3.0
smatch==1.0.4
smmap==5.0.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.2
srsly==2.4.5
stanza==1.5.0
stevedore==4.1.1
supar==1.1.4
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
thinc==8.1.7
tokenizers==0.13.2
torch==1.12.1
torchaudio==0.12.1
torchdiffeq==0.2.3
torchmetrics==0.11.0
torchvision==0.13.1
transformers==4.26.0
tree-sitter==0.20.1
typer==0.7.0
Unidecode==1.3.6
uri-template==1.2.0
virtualenv==20.18.0
wandb==0.14.0
wasabi==1.1.1
webcolors==1.12
webencodings==0.5.1
Werkzeug==2.2.3
x-transformers==1.10.0
xxhash==3.2.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.12.1
